\docType{package}
\name{lambda.tools-package}
\alias{lambda.tools}
\alias{lambda.tools-package}
\title{Tools for functional programming in R}
\description{
  This package contains a collection of functions that
  facilitate modeling of data using a functional
  programming paradigm. The idea is that using tools that
  are more closely connected with the idioms of mathematics
  will make it easier to map the mathematical model to the
  software model.
}
\details{
  \tabular{ll}{ Package: \tab lambda.tools\cr Type: \tab
  Package\cr Version: \tab 1.0.0\cr Date: \tab
  2013-08-30\cr License: \tab LGPL-3\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Brian Lee Yung Rowe <r@zatonovo.com>
}
\seealso{
  \code{\link{lambda.r}}
}
\keyword{attribute}
\keyword{logic}
\keyword{package}

